completata ristrutturazione utils.ts
